fix: update project configuration and improve error handling
- Added new entries to `.gitignore` for `.agents/`, `.claude/`, and `.cursor/` directories.
- Enhanced `Cargo.toml` with linting configurations for Rust and Clippy to enforce code quality.
- Refactored error handling in `main.rs` to use `context` for clearer error messages.
- Introduced a helper method `handle_request` in `RublClient` to streamline request handling and reduce code duplication across multiple endpoints.
- Updated data types in `FetchNearbyHotspotsRequest` and related structures from `f32` to `f64` for improved precision.
- Modified formatting functions in `trip.rs` to use `unwrap_or_else` for better performance and readability.

+++ src/tools/hotspot.rs
@@ -71,11 +71,11 @@ impl Endpoint for FetchRegionHotspotsRequest {
 #[derive(Debug, Serialize, Deserialize, schemars::JsonSchema)]
 pub struct FetchNearbyHotspotsRequest {
     #[schemars(description = "Latitude")]
-    pub lat: f32,
+    pub lat: f64,
     #[schemars(description = "Longitude")]
-    pub lng: f32,
+    pub lng: f64,
     #[schemars(description = "Radius in kilometers", range(min = 0, max = 500))]
-    pub dist: Option<f32>,
+    pub dist: Option<f64>,
     #[schemars(
         description = "Only fetch hotspots visited up to back days ago",
         range(min = 1, max = 30)
